+++ packages/shared/src/hooks/index.ts
@@ -5,3 +5,6 @@ export { useIsMounted } from './useIsMounted';
 export const hooks = {
     useCustomWebkitScrollbar,
 };
+
+export * from './use-force-update';
+export * from './use-ref-as-state';