+++ packages/calendar/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @alfalab/core-components-calendar
 
+## 9.1.7
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-typography@6.0.6
+- @alfalab/core-components-modal@11.0.11
+
+## 9.1.6
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.0
+- @alfalab/core-components-button@13.1.3
+- @alfalab/core-components-modal@11.0.10
+- @alfalab/core-components-mq@6.0.4
+- @alfalab/core-components-typography@6.0.5
+- @alfalab/core-components-icon-button@8.0.6
+
+## 9.1.5
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-button@13.1.2
+- @alfalab/core-components-icon-button@8.0.5
+- @alfalab/core-components-modal@11.0.9
+
+## 9.1.4
+
+### Patch Changes
+
+<sup><time>08.05.2026</time></sup>
+
+#### [#2190](https://github.com/core-ds/core-components/pull/2190)
+
+##### Calendar
+
+- Исправлен `border-radius` кнопок выбора месяца и года
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-modal@11.0.8
+
 ## 9.1.3
 
 ### Patch Changes