+++ packages/circular-progress-bar/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @alfalab/core-components-circular-progress-bar
 
+## 5.0.6
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-typography@6.0.6
+
+## 5.0.5
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.0
+- @alfalab/core-components-typography@6.0.5
+
 ## 5.0.4
 
 ### Patch Changes