+++ packages/intl-phone-input/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @alfalab/core-components-intl-phone-input
 
+## 11.0.15
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-select@19.1.8
+- @alfalab/core-components-input-autocomplete@14.0.15
+
+## 11.0.14
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-input-autocomplete@14.0.14
+- @alfalab/core-components-select@19.1.7
+
+## 11.0.13
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-select@19.1.6
+- @alfalab/core-components-input-autocomplete@14.0.13
+
+## 11.0.12
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-select@19.1.5
+- @alfalab/core-components-input-autocomplete@14.0.12
+
 ## 11.0.11
 
 ### Patch Changes