+++ packages/list/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alfalab/core-components-list
 
+## 5.0.6
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-typography@6.0.6
+
+## 5.0.5
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-typography@6.0.5
+
 ## 5.0.4
 
 ### Patch Changes