+++ packages/masked-input/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @alfalab/core-components-masked-input
 
+## 8.0.8
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-input@17.1.6
+
+## 8.0.7
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-input@17.1.5
+
+## 8.0.6
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-input@17.1.4
+
 ## 8.0.5
 
 ### Patch Changes