+++ packages/navigation-bar-private/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @alfalab/core-components-navigation-bar-private
 
+## 2.0.9
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-typography@6.0.6
+
+## 2.0.8
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.0
+- @alfalab/core-components-button@13.1.3
+- @alfalab/core-components-typography@6.0.5
+- @alfalab/core-components-icon-button@8.0.6
+
+## 2.0.7
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-button@13.1.2
+- @alfalab/core-components-icon-button@8.0.5
+
 ## 2.0.6
 
 ### Patch Changes