+++ packages/typography/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @alfalab/core-components-typography
 
+## 6.0.6
+
+### Patch Changes
+
+<sup><time>29.05.2026</time></sup>
+
+#### [#2226](https://github.com/core-ds/core-components/pull/2226)
+
+##### Typography
+
+- Для `alfasans` версии пакета оставлен только один вариант шрифта, проп `font` игнорируется
+
+## 6.0.5
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.0
+- @alfalab/core-components-mq@6.0.4
+
 ## 6.0.4
 
 ### Patch Changes